+++ R/verb-distinct.R
@@ -19,10 +19,21 @@ distinct.tbl_lazy <- function(.data, ..., .keep_all = FALSE) {
   empty_dots <- dots_n(...) == 0
   can_use_distinct <- !.keep_all || (empty_dots && is_empty(grps))
   if (!can_use_distinct) {
+    needs_dummy_order <- is.null(op_sort(.data))
+
+    if (needs_dummy_order) {
+      dummy_order_vars <- colnames(.data)[[1]]
+      .data <- .data %>% window_order(!!sym(dummy_order_vars))
+    }
+
     .data <- .data %>%
       group_by(..., .add = TRUE) %>%
       filter(row_number() == 1L) %>%
       group_by(!!!grps)
+
+    if (needs_dummy_order) {
+      .data <- .data %>% window_order()
+    }
 
     return(.data)
   }